4_asyncio/requirements.txt
asyncio==3.4.3
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.1
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
xlrd==2.0.1
